+++ packages/core/src/components/AttentionBox/__stories__/AttentionBox.stories.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import { useCallback } from "react";
 import AttentionBox from "../AttentionBox";
 import { createStoryMetaSettingsDecorator } from "../../../storybook";
@@ -12,6 +12,7 @@ import person from "./assets/person.png";
 import Flex from "../../Flex/Flex";
 import AttentionBoxLink from "../AttentionBoxLink/AttentionBoxLink";
 import "./AttentionBox.stories.scss";
+import Button from "../../Button/Button";
 
 const metaSettings = createStoryMetaSettingsDecorator({
   component: AttentionBox,
@@ -174,6 +175,41 @@ export const AttentionBoxInsideADialogCombobox = {
           <AttentionBox text="Hold ⌘ to select more than one person or team" compact onClose={mockOnClose} />
         </div>
       </DialogContentContainer>
+    );
+  },
+
+  name: "Attention box inside a dialog/combobox"
+};
+
+export const AttentionBoxAnimation = {
+  render: () => {
+    const [isOpen, setOpen] = useState(false);
+    const onClick = useCallback(() => {
+      setOpen(true);
+    }, []);
+    return (
+      <>
+        <Flex
+          gap={Flex.gaps.MEDIUM}
+          style={{
+            height: "44px"
+          }}
+        >
+          <Button onClick={onClick} kind={Button.kinds.SECONDARY}>
+            Entry animation
+          </Button>
+          {isOpen && (
+            <AttentionBox
+              compact
+              withIconWithoutHeader
+              entryAnimation
+              icon={Info}
+              text="First, move the content you want to copy into folder. Only main boards and dashboards can be copied."
+              onClose={() => setOpen(false)}
+            />
+          )}
+        </Flex>
+      </>
     );
   },