+++ packages/core/src/components/Table/TableHeaderCell/TableHeaderCell.tsx
@@ -14,7 +14,7 @@ import { ComponentDefaultTestId } from "../../../tests/constants";
 import { getStyle } from "../../../helpers/typesciptCssModulesHelper";
 
 export interface TableHeaderCellProps extends VibeComponentProps {
-  title: string;
+  title: string | React.ReactNode;
   icon?: SubIcon;
   infoContent?: string;
   sortState?: "asc" | "desc" | "none";
@@ -64,11 +64,13 @@ const TableHeaderCell: VibeComponent<TableHeaderCellProps, HTMLDivElement> = for
       >
         <Flex direction="row" align="center" className={styles.tableHeaderCellContent} gap="xs">
           {icon && <Icon icon={icon} className={styles.icon} />}
-          {
+          {typeof title === "string" ? (
             <Text type="text2" weight="medium" color="secondary">
               {title}
             </Text>
-          }
+          ) : (
+            title
+          )}
           {infoContent && (
             <Tooltip content={infoContent} referenceWrapperClassName={styles.infoTooltip}>
               <Icon icon={Info} iconLabel={infoContent} />