+++ packages/core/src/components/Table/context/TableContainerContext/TableContainerContext.types.ts
@@ -1,9 +1,12 @@
 import React from "react";
 
-export interface TableContainerContext {
-  menuContainerRef: React.MutableRefObject<HTMLDivElement>;
+export type TableContainerContext = TableContainerProviderValue;
+
+export interface TableContainerProviderValue {
+  menuContainerRef: React.RefObject<HTMLDivElement>;
 }
 
 export interface TableContainerProviderProps {
+  value: TableContainerProviderValue;
   children: React.ReactNode;
 }