+++ packages/icons/src/iconsMetaData.ts
@@ -14,6 +14,20 @@ export default [
   // plop_marker:icon_metadata
 
   {
+    name: "Clipboard",
+    file: "Clipboard.svg",
+    description: "Clipboard",
+    tags: "Clipboard, Board, Copy"
+  },
+
+  {
+    name: "Forward",
+    file: "Forward.svg",
+    description: "Forward",
+    tags: "Forward, Email, Send, Arrow, Right"
+  },
+
+  {
     name: "Shuffle",
     file: "Shuffle.svg",
     description: "Use to describe shuffle of items.",
@@ -1802,7 +1816,7 @@ export default [
     name: "Wand",
     file: "Wand.svg",
     description: "Wand",
-    tags: "Wand, Star",
+    tags: "Wand, Star, Template",
     category: [PLATFORM]
   },